src/tax-system/requirements.txt
# Core
pandas>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0

# Database
# SQLite is built-in to Python

# Excel output
openpyxl>=3.1.0

# OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# Email parsing
# imaplib and email are built-in to Python

# Browser automation (screenshots / PDF)
playwright>=1.40.0

# Fuzzy matching
rapidfuzz>=3.0.0

# Currency conversion
requests>=2.31.0

# CLI
click>=8.1.0
rich>=13.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
